backend/requirements.txt
fastapi>=0.104.1
uvicorn>=0.24.0
python-multipart>=0.0.6
python-pptx>=0.6.23
pandas>=2.2.2
openpyxl>=3.1.2
xlrd>=2.0.1
fuzzywuzzy>=0.18.0
python-levenshtein>=0.23.0
openai>=1.3.7
reportlab>=4.0.7
Pillow>=10.1.0
python-docx>=1.1.0
requests>=2.31.0
pydantic>=2.5.0